fix(swift-sdk)!: realign Swift FFI shims with current rust-dashcore + platform-wallet-ffi
After bumping rust-dashcore to feat/platform-wallet2 rev 4c8bec36 and
resolving the v3.1-dev merge, the Swift layer had 18 compile errors
against the refreshed C ABI. Fixes them plus a handful of related
drift items that only surfaced once the first batch cleared.

**Umbrella header / module map (packages/swift-sdk/build_ios.sh)**
- dash-network.h comes first (defines FFINetwork referenced by everyone).
- dashcore/dashcore.h excluded via `exclude header` — it duplicates
  FFINetwork with dash-network.h and triggers typedef-redefinition
  (dash-network was extracted as a standalone crate upstream).
- platform-wallet-ffi headers stay excluded: Swift uses @_silgen_name
  shims in PlatformWalletFFI.swift as the source of truth.

**@_silgen_name shim realignment (PlatformWalletFFI.swift)**
- platform_wallet_info_create_from_{seed,mnemonic}: added `network: u32`
  first arg (see sibling commit making the Rust side C-ABI clean).
- platform_wallet_info_{get,set}_identity_manager: dropped network arg.
- contact_request_create: added missing `core_height_created_at: u32`.
- managed_identity_send/accept_contact_request: reduced to
  (identity_handle, request_handle, out_error).
- platform_wallet_bytes_free: added `len` param.
- *_destroy shims (contact_request, established_contact, identity_manager,
  managed_identity, platform_wallet_info): return PlatformWalletFFIResult
  now, not Void.

**Swift call-site adjustments**
- ContactRequest.swift / IdentityManager.swift: Int for usize-typed lens.
- ManagedIdentity.swift: dropped UInt wrapper; renamed stale
  PlatformBlockTime references to BlockTime / FFIBlockTime.
- ManagedPlatformWallet.swift: let ffiInputs (never-mutated).
- PlatformWallet.swift: drop UInt(seed.count) wrapper.

**KeyWalletTypes.swift**
- Deleted the duplicate public `enum TransactionContext: UInt32` that
  collided with the `TransactionContext` class in TransactionContext.swift.
- TransactionContextDetails.context now typed as TransactionContextType.
- Transaction.swift / ManagedWallet.swift: follow the rename in their
  default arg types.

**Swift 6.2 compat (ConcurrencyCompat.swift)**
- Swift 6.2 stdlib restates OpaquePointer's Sendable conformance as
  `@available(*, unavailable)`, which makes our retroactive
  `extension OpaquePointer: @unchecked Sendable` trip
  "already stated in type's module 'Swift'" under -warnings-as-errors,
  while simultaneously blocking bare-OpaquePointer round-trips through
  concurrency boundaries.
- Gate the retroactive shim behind `#if compiler(<6.2)` and add a
  SendableOpaquePointer wrapper for 6.2 call sites.
- TransitionDetailView.swift: all 14 `try await MainActor.run { ... } ->
  OpaquePointer` sites now wrap/unwrap SendableOpaquePointer.

**Dead-code cleanup**
- Deleted AddAccountView.swift — referenced the already-removed
  WalletService / HDWallet types (commit 6ee01f9ce), only surviving
  because the Xcode project uses PBXFileSystemSynchronizedRootGroup.
- ReceiveAddressView.swift: unifiedAppState.platformState → platformState
  (UnifiedAppState removed in the iOS SDK refactor).

**Rust side (packages/rs-platform-wallet-ffi/)**
- Dropped the earlier build.rs + cbindgen build-dep: now that Swift is
  not consuming the generated header, the cbindgen emit is dead work.
- Kept the earlier Network-by-value → u32 signature fix on the two
  entry points (independently correct; matches rest of the crate).

Verification:
- `bash packages/swift-sdk/build_ios.sh --target sim --profile dev` → clean
- `cargo clippy --workspace --all-features --locked --all-targets --
  --no-deps -D warnings` → 0 errors
- `cargo machete` → no unused deps

Pushed by sibling commits; this one closes the loop on CI red.

+++ packages/swift-sdk/Sources/SwiftDashSDK/PlatformWallet/ContactRequest.swift
@@ -42,7 +42,7 @@ public final class ContactRequest: @unchecked Sendable {
                 recipientKeyIndex,
                 accountReference,
                 keyPtr.baseAddress?.assumingMemoryBound(to: UInt8.self),
-                UInt(encryptedPublicKey.count),
+                encryptedPublicKey.count,
                 coreHeightCreatedAt,
                 createdAt,
                 &handle,
@@ -125,7 +125,7 @@ public final class ContactRequest: @unchecked Sendable {
     /// Get the encrypted public key
     public func getEncryptedPublicKey() throws -> Data {
         var bytesPtr: UnsafeMutablePointer<UInt8>? = nil
-        var length: UInt = 0
+        var length: Int = 0
         var error = PlatformWalletFFIError()
 
         let result = contact_request_get_encrypted_public_key(handle, &bytesPtr, &length, &error)
